rename auth form files to login and register
