+++ src/sections/auth/login/index.js
@@ -1,1 +1,1 @@
-export { default as LoginForm } from './LoginForm';
+export { default as LoginForm } from './Login';
