+++ src/sections/auth/register/index.js
@@ -1,1 +1,1 @@
-export { default as RegisterForm } from './RegisterForm';
+export { default as RegisterForm } from './Register';
